Move selector functions to data namespace
