+++ frontend/src/metabase/nav/containers/MainNavbar/MainNavbarContainer/MainNavbarContainer.tsx
@@ -20,7 +20,7 @@ import Collections, {
 } from "metabase/entities/collections";
 import { logout } from "metabase/auth/actions";
 import { getUser, getUserIsAdmin } from "metabase/selectors/user";
-import { getHasDataAccess, getHasOwnDatabase } from "metabase/nav/selectors";
+import { getHasDataAccess, getHasOwnDatabase } from "metabase/selectors/data";
 
 import CollectionCreate from "metabase/collections/containers/CollectionCreate";
 import {
